requirements.txt
fastapi
uvicorn[standard]
anthropic
pyjwt
bcrypt
